requirements.txt
Django==4.2.6

django-filter~=23.3